Installer: drop the lemonade ROCm fallback now the fork ships identical per-gfx prebuilts
