+++ studio/setup.ps1
@@ -850,7 +850,7 @@ if (-not $HasNvidiaSmi) {
     # popping a UAC/DiskPart prompt RunAsInvoker can't suppress (its manifest is
     # asInvoker; even 'amd-smi version' hangs). So only probe when a HIP SDK is present
     # (hipinfo found -> un-elevated) or the user opts in; else fall through to WMI name
-    # inference (enough to pick ROCm wheels + lemonade llama.cpp).
+    # inference (enough to pick ROCm wheels + the ROCm llama.cpp prebuilt).
     # An explicit opt-out (UNSLOTH_ENABLE_AMD_SMI=0/false/no/off) wins over the HIP-SDK
     # heuristic: a HIP SDK binary with a broken runtime can still pop the prompt, so
     # $HipSdkInstalled must NOT silently re-enable it.
@@ -923,7 +923,7 @@ if (-not $HasNvidiaSmi) {
     }
     # ── Arch resolution: env-var override → name inference ──────────────────
     # Runs after all probes, even when none confirmed a ROCm runtime ($HasROCm false):
-    # the Adrenalin driver alone runs the lemonade-sdk llama.cpp prebuilt (bundles its
+    # the Adrenalin driver alone runs the per-gfx ROCm llama.cpp prebuilt (bundles its
     # own runtime), and all it needs is the gfx arch, inferable from the WMI GPU name.
     # Resolving it here lets setup.ps1 forward --rocm-gfx so a GPU llama.cpp is pulled
     # instead of CPU. (PyTorch ROCm wheels still require a HIP SDK -- gated on $HasROCm
@@ -936,7 +936,7 @@ if (-not $HasNvidiaSmi) {
             substep "gfx arch from UNSLOTH_ROCM_GFX_ARCH env override: $script:ROCmGfxArch" "Cyan"
         }
         # 2. Best-effort name → arch lookup (amd-smi / WMI). Most-specific first,
-        #    first match wins. Covers only arches the lemonade-sdk prebuilts support
+        #    first match wins. Covers only arches the ROCm prebuilts support
         #    (gfx120X/110X/1151/1150/103X); unknown names fall back cleanly to CPU.
         elseif ($ROCmGpuLabel) {
             $nameArchTable = @(
@@ -947,9 +947,9 @@ if (-not $HasNvidiaSmi) {
                 @{ P = "RX 7900|RX 7800|RX 7700(?!S)|PRO W7900|PRO W7800|PRO W7700"; A = "gfx1100" }  # RDNA 3 desktop / workstation (Navi 31)
                 @{ P = "RX 7600|RX 7700S|RX 7650|PRO W7600|PRO W7500|PRO V710"; A = "gfx1102" }  # RDNA 3 (Navi 33)
                 @{ P = "780M|760M|740M|Phoenix|Hawk Point|Z1 Extreme|Z2 Extreme"; A = "gfx1103" }  # RDNA 3 iGPU (Phoenix / Hawk Point)
-                @{ P = "RX 6900|RX 6800|RX 6750|RX 6700|PRO W6800|PRO W6900";  A = "gfx1030" }  # RDNA 2 (Navi 21) -- lemonade gfx103X
-                @{ P = "RX 6650|RX 6600|PRO W6600|PRO W6650";                  A = "gfx1032" }  # RDNA 2 (Navi 23) -- lemonade gfx103X
-                @{ P = "RX 6500|RX 6400|RX 6300|PRO W6400|PRO W6500";          A = "gfx1034" }  # RDNA 2 (Navi 24) -- lemonade gfx103X
+                @{ P = "RX 6900|RX 6800|RX 6750|RX 6700|PRO W6800|PRO W6900";  A = "gfx1030" }  # RDNA 2 (Navi 21) -- gfx103X family
+                @{ P = "RX 6650|RX 6600|PRO W6600|PRO W6650";                  A = "gfx1032" }  # RDNA 2 (Navi 23) -- gfx103X family
+                @{ P = "RX 6500|RX 6400|RX 6300|PRO W6400|PRO W6500";          A = "gfx1034" }  # RDNA 2 (Navi 24) -- gfx103X family
             )
             foreach ($row in $nameArchTable) {
                 if ($ROCmGpuLabel -match $row.P) {
@@ -2754,7 +2754,7 @@ if ($env:UNSLOTH_LLAMA_FORCE_COMPILE -eq "1") {
                 # or the upstream windows-hip fallback, so accept either and never
                 # treat a valid ROCm install as mismatched. A name-inferred gfx
                 # arch (Adrenalin-only, no confirmed runtime) still counts as
-                # ROCm-capable -- the lemonade prebuilt bundles its own runtime,
+                # ROCm-capable -- the ROCm prebuilt bundles its own runtime,
                 # mirroring the --rocm-gfx forward below. NOTE: this block is
                 # currently inert -- write_prebuilt_metadata does not persist an
                 # install_kind key, so $existingKind is always null. If that changes,
@@ -2785,7 +2785,7 @@ if ($env:UNSLOTH_LLAMA_FORCE_COMPILE -eq "1") {
         if ($HasROCm) {
             $prebuiltArgs += "--has-rocm"
         }
-        # Forward the resolved gfx arch so the lemonade HIP prebuilt is picked even
+        # Forward the resolved gfx arch so the per-gfx ROCm prebuilt is picked even
         # when the installer's probe can't confirm the runtime (amd-smi-only /
         # Adrenalin-only, name-inferred arch). --rocm-gfx is authoritative and
         # implies ROCm in install_llama_prebuilt.py, so the GPU prebuilt is selected
@@ -2971,10 +2971,10 @@ if (-not $NeedLlamaSourceBuild) {
     } elseif ($HasROCm -or $script:ROCmGfxArch) {
         # AMD GPU present but in the CPU-only source-build fallback: a HIP source
         # build needs the full HIP SDK + ROCm clang toolchain. AMD GPU acceleration
-        # comes from the lemonade prebuilt (bundles the runtime, no SDK) -- reaching
+        # comes from the per-gfx ROCm prebuilt (bundles the runtime, no SDK) -- reaching
         # here means it couldn't be installed. Warn loudly, don't ship a slow CPU build.
         $_amdArch = if ($script:ROCmGfxArch) { $script:ROCmGfxArch } else { "ROCm" }
-        substep "[WARN] AMD GPU ($_amdArch) detected, but the GPU-accelerated lemonade" "Yellow"
+        substep "[WARN] AMD GPU ($_amdArch) detected, but the GPU-accelerated ROCm" "Yellow"
         substep "       llama.cpp prebuilt could not be installed -- falling back to a CPU build." "Yellow"
         substep "       The prebuilt is the AMD GPU path (no HIP SDK required). To restore GPU" "Yellow"
         substep "       acceleration: re-run the installer (check your network / proxy), or set" "Yellow"
